+++ src/components/CardList.js
@@ -1,23 +1,5 @@
-import timeAgo from "./util/timeAgo.js";
-import date from "./util/date.js";
-import baseImg from "../assets/png/base.png";
-
-function CardListItem({ item }) {
-  const formatDate = date(item.createdAt);
-
-  return (
-    <a className="card_list_item" href={item.url} target="_blank" rel="noreferrer noopener">
-      <div className="img_area">
-        <img src={item.imageSource || baseImg} alt="" />
-      </div>
-      <div className="text_area">
-        <p className="time">{timeAgo(formatDate)}</p>
-        <p className="description">{item.description}</p>
-        <p className="created_at">{formatDate}</p>
-      </div>
-    </a>
-  );
-}
+import CardListItem from "./CardListItem";
+import "./CardList.css";
 
 function CardList({ items }) {
   return (
