+++ src/components/CardsArea.css
@@ -41,71 +41,6 @@
   line-height: 1.5; /* 150% */
 }
 
-.card_list {
-  display: flex;
-  flex-wrap: wrap;
-  gap: 25px 20px;
-}
-
-.card_list_item {
-  width: 340px;
-  height: 100%;
-  display: flex;
-  flex-direction: column;
-  border-radius: 15px;
-  box-shadow: 0px 5px 25px 0px rgba(0, 0, 0, 0.08);
-  overflow: hidden;
-  cursor: pointer;
-  text-decoration-line: none;
-}
-
-.img_area {
-  overflow: hidden;
-}
-
-.img_area img {
-  width: 100%;
-  height: 200px;
-  flex-shrink: 0;
-  transition: transform 0.3s ease;
-  object-fit: cover;
-}
-
-.card_list_item:hover .img_area img {
-  transform: scale(1.3);
-}
-
-.text_area {
-  display: flex;
-  padding: 15px 20px;
-  flex-direction: column;
-  align-items: flex-start;
-
-  gap: 10px;
-}
-
-.time {
-  color: #666;
-  font-size: 1.3rem;
-  font-weight: 400;
-}
-
-.description {
-  max-height: 50px;
-  overflow: hidden;
-  color: #000;
-  text-overflow: ellipsis;
-  font-size: 1.6rem;
-  font-weight: 400;
-  line-height: 1.5;
-}
-
-.created_at {
-  color: #333;
-  font-size: 1.4rem;
-  font-weight: 400;
-}
-
 @media (max-width: 1124px) {
   .card_inner {
     width: auto;
